slang_normalizer/pyproject.toml
[project]
name = "slang-normalizer"
version = "0.1.0"
description = "A research package for slang normalization with local MLX fine-tuning and evaluation pipelines."
readme = "README.md"
authors = [
    { name = "wz335", email = "wz335@georgetown.edu" }
]
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.8.4",
    "mlx-lm[train]>=0.31.2",
    "openai>=2.32.0",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "torch>=2.11.0",
    "transformers>=5.5.4",
]

[build-system]
requires = ["uv_build>=0.11.6,<0.12.0"]
build-backend = "uv_build"

[project.scripts]
slang-normalizer-preprocess = "slang_normalizer.preprocess:main"
slang-normalizer-prepare-mlx = "slang_normalizer.prepare_mlx_data:main"
slang-normalizer-train-mlx = "slang_normalizer.train_mlx:main"
slang-normalizer-download-metadata = "slang_normalizer.download_benchmark_metadata:main"
slang-normalizer-baseline = "slang_normalizer.baseline_eval:run"
slang-normalizer-finetuned = "slang_normalizer.finetuned_inference:main"
slang-normalizer-advanced = "slang_normalizer.advanced_with_backtranslation:main"
slang-normalizer-debate = "slang_normalizer.multi_agent_debate:run"
slang-normalizer-demo = "slang_normalizer.demo_advanced_backtranslation:main"
slang-normalizer-prepare-dpo = "slang_normalizer.prepare_dpo_data:main"
slang-normalizer-train-dpo = "slang_normalizer.train_dpo_mlx:main"
slang-normalizer-combine = "slang_normalizer.combine_results:main"

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.3",
    "ruff>=0.15.11",
]

[tool.ruff]
line-length = 88
target-version = "py313"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
